vchan: Fix handling of invalid ring buffer indices
The remote (hostile) process can set ring buffer indices to any value at
any time. If that happens, it is possible to get "buffer space" (either
for writing data, or ready for reading) negative or greater than
buffer size. This will end up with buffer overflow in _write/_read
function because of not strong enough check for "size" parameter -
compared then to malicious "avail" value.

+++ vchan/libvchan.h
@@ -54,8 +54,8 @@ struct libvchan *libvchan_server_init(int devno);
 
 struct libvchan *libvchan_client_init(int domain, int devno);
 
-int libvchan_write(struct libvchan *ctrl, char *data, int size);
-int libvchan_read(struct libvchan *ctrl, char *data, int size);
+int libvchan_write(struct libvchan *ctrl, char *data, size_t size);
+int libvchan_read(struct libvchan *ctrl, char *data, size_t size);
 int libvchan_wait(struct libvchan *ctrl);
 int libvchan_close(struct libvchan *ctrl);
 void libvchan_prepare_to_select(struct libvchan *ctrl);
